notifications.rs
Both get and post functions are ready for testing.
Get: is a get all, as there in never a reason to only grab one singular part of the object, as without a table the details are useless and vice versa.

Post: creates notifications of the form { (table : int), (notif_type : int), (details : string) }, in which table is the unique table identifier, notif_type corresponds to an enum we will establish later and details is useful information (i.e. drinks required for a refill notification).